+++ CollisionHandler.cpp
@@ -11,7 +11,7 @@ bool CollisionHandler::BoxToBoxCollision(BoundingBox a, BoundingBox b)
 
 bool CollisionHandler::SphereToSphereCollision(BoundingSphere a, BoundingSphere b)
 { 
-    float distance = sqrt((XMVectorGetX(a.centreWorldPos) - XMVectorGetX(b.centreWorldPos)) * (XMVectorGetX(a.centreWorldPos) - XMVectorGetX(b.centreWorldPos)) +
+    float distance = (float)sqrt((XMVectorGetX(a.centreWorldPos) - XMVectorGetX(b.centreWorldPos)) * (XMVectorGetX(a.centreWorldPos) - XMVectorGetX(b.centreWorldPos)) +
         (XMVectorGetY(a.centreWorldPos) - XMVectorGetY(b.centreWorldPos)) * (XMVectorGetY(a.centreWorldPos) - XMVectorGetY(b.centreWorldPos)) +
         (XMVectorGetZ(a.centreWorldPos) - XMVectorGetZ(b.centreWorldPos)) * (XMVectorGetZ(a.centreWorldPos) - XMVectorGetZ(b.centreWorldPos)));
     
@@ -25,7 +25,7 @@ bool CollisionHandler::SphereToBoxCollision(BoundingSphere a, BoundingBox b)
     float y = std::max(XMVectorGetY(b.minBoundV), std::min(XMVectorGetY(a.centreWorldPos), XMVectorGetY(b.maxBoundV)));
     float z = std::max(XMVectorGetZ(b.minBoundV), std::min(XMVectorGetZ(a.centreWorldPos), XMVectorGetZ(b.maxBoundV)));
      
-    float distance = sqrt((x - XMVectorGetX(a.centreWorldPos)) * (x - XMVectorGetX(a.centreWorldPos)) +
+    float distance = (float)sqrt((x - XMVectorGetX(a.centreWorldPos)) * (x - XMVectorGetX(a.centreWorldPos)) +
         (y - XMVectorGetY(a.centreWorldPos)) * (y - XMVectorGetY(a.centreWorldPos)) +
         (z - XMVectorGetZ(a.centreWorldPos)) * (z - XMVectorGetZ(a.centreWorldPos)));
